Consolidating code
I prefer code to be as concise as possible.  Also each function to be commented with instructions as to what it is doing.  Code should not be repeated too many times.  I've just made the above changes so you can see how I like the code to be done but I haven't tested it.  

Please test it and modify it such that it works.  The above should give you an idea of the quality of codework that is expected.